Develop

+++ customer/views.py
@@ -8,8 +8,9 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 
-from customer.serializers import CustomerSerializer
-from customer.models import Customer
+from customer.serializers import CustomerSerializer, AddressSerializer
+from customer.models import Customer, Address
+from customer.permissions import IsOwnerOrReadOnly
 
 # It seems that two class based views is the best option
 class CustomerList(APIView):
@@ -39,6 +40,8 @@ class CustomerDetail(APIView):
     """
         Retrieve, update or delete a customer instance.
     """
+    permission_classes = ( IsOwnerOrReadOnly,)
+
     def get_object(self, pk):
         try:
             return Customer.objects.get(pk = pk)
@@ -47,11 +50,13 @@ class CustomerDetail(APIView):
 
     def get(self, request, pk, format = None):
         customer = self.get_object(pk)
+        self.check_object_permissions(self.request, customer)
         serializer = CustomerSerializer(customer)
         return Response(serializer.data)
 
     def put(self, request, pk, format = None):
         customer = self.get_object(pk)
+        self.check_object_permissions(self.request, customer)
         serializer = CustomerSerializer(data=request.data)
         if serializer.is_valid():
             user = customer.user
@@ -66,6 +71,80 @@ class CustomerDetail(APIView):
 
     def delete(self, request, pk, format = None):
         customer = self.get_object(pk)
+        self.check_object_permissions(self.request, customer)
         customer.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
+class AddressList(APIView):
+    """
+        List all customers or create new customer endpoint
+    """
+    def get(self, request, format = None):
+        addresses = Address.objects.all()
+        serializer = AddressSerializer(addresses, many = True)
+        return Response(serializer.data)
+
+    def post(self, request, format = None):
+        serializer = AddressSerializer(data = request.data)
+        if serializer.is_valid():
+            customer = Customer.objects.filter(user__id__iexact=request.user.id).first()
+            address = Address(name = serializer.data['name'], \
+                        address = serializer.data['address'], \
+                        city = serializer.data['city'],       \
+                        country = serializer.data['country'], \
+                        customer = customer)
+            address.save()
+            return Response(serializer.data, status = status.HTTP_201_CREATED)
+        return Response(serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+
+
+class AddressDetail(APIView):
+    """
+        Retrieve, update or delete a customer instance.
+    """
+    permission_classes = ( IsOwnerOrReadOnly,)
+
+    def get_object(self, pk):
+        try:
+            return Address.objects.get(pk = pk)
+        except Address.DoesNotExist:
+            raise Http404
+
+    def get(self, request, pk, format = None):
+        address = self.get_object(pk)
+        self.check_object_permissions(self.request, address)
+        serializer = AddressSerializer(address)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format = None):
+        address = self.get_object(pk)
+        self.check_object_permissions(self.request, address)
+        serializer = AddressSerializer(data=request.data)
+        if serializer.is_valid():
+            address.name = serializer.data['name']
+            address.address = serializer.data['address']
+            address.city = serializer.data['city']
+            address.country = serializer.data['country']
+            address.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format = None):
+        address = self.get_object(pk)
+        self.check_object_permissions(self.request, address)
+        address.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+@api_view(['GET'])
+def get_customer_authenticated(request):
+    """
+    Retrieve customer instance given the auth token.
+    """
+    try:
+        user = request.user
+        customer = Customer.objects.filter(user__id__iexact=user.id).first()
+    except Snippet.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+        
+    serializer = CustomerSerializer(customer)
+    return Response(serializer.data)
